refactor(chat): implement SWR for chat config data fetching
Replace manual useEffect/useState pattern with SWR for improved:
- Automatic caching and deduplication
- Background revalidation control
- Better error handling and loading states
- Reduced duplicate API calls with 5-minute deduping interval

Changes:
- Add SWR dependency for data fetching
- Refactor Thread component to use useSWR hook
- Rename chat.ts to chatService.ts for consistency
- Configure fallback data for offline/error scenarios
- Remove manual loading state and effect management

The SWR implementation provides more robust data fetching with
built-in caching, automatic revalidation, and better UX during
network failures.
