+++ package.json
@@ -44,6 +44,7 @@
     "react-use-websocket": "3.0.0",
     "remark-gfm": "^4.0.1",
     "streamdown": "^1.2.0",
+    "swr": "^2.3.6",
     "tailwind-merge": "^3.3.1",
     "three": "^0.180.0",
     "tokenlens": "^1.2.1",
